+++ packages/discord.js/src/client/websocket/handlers/AUTO_MODERATION_ACTION_EXECUTION.js
@@ -1,5 +1,17 @@
 'use strict';
 
-module.exports = (client, packet) => {
-  client.actions.AutoModerationActionExecution.handle(packet.d);
+const AutoModerationActionExecution = require('../../../structures/AutoModerationActionExecution.js');
+const Events = require('../../../util/Events.js');
+
+module.exports = (client, { d: data }) => {
+  const guild = client.guilds.cache.get(data.guild_id);
+  if (!guild) return;
+
+  /**
+   * Emitted whenever an auto moderation rule is triggered.
+   * <info>This event requires the {@link PermissionFlagsBits.ManageGuild} permission.</info>
+   * @event Client#autoModerationActionExecution
+   * @param {AutoModerationActionExecution} autoModerationActionExecution The data of the execution
+   */
+  client.emit(Events.AutoModerationActionExecution, new AutoModerationActionExecution(data, guild));
 };
