+++ packages/discord.js/src/client/websocket/handlers/GUILD_AUDIT_LOG_ENTRY_CREATE.js
@@ -1,5 +1,19 @@
 'use strict';
 
-module.exports = (client, packet) => {
-  client.actions.GuildAuditLogEntryCreate.handle(packet.d);
+const GuildAuditLogsEntry = require('../../../structures/GuildAuditLogsEntry.js');
+const Events = require('../../../util/Events.js');
+
+module.exports = (client, { d: data }) => {
+  const guild = client.guilds.cache.get(data.guild_id);
+  if (!guild) return;
+
+  const auditLogEntry = new GuildAuditLogsEntry(guild, data);
+
+  /**
+   * Emitted whenever a guild audit log entry is created.
+   * @event Client#guildAuditLogEntryCreate
+   * @param {GuildAuditLogsEntry} auditLogEntry The entry that was created
+   * @param {Guild} guild The guild where the entry was created
+   */
+  client.emit(Events.GuildAuditLogEntryCreate, auditLogEntry, guild);
 };
