+++ packages/discord.js/src/client/websocket/handlers/GUILD_SCHEDULED_EVENT_CREATE.js
@@ -1,5 +1,17 @@
 'use strict';
 
-module.exports = (client, packet) => {
-  client.actions.GuildScheduledEventCreate.handle(packet.d);
+const Events = require('../../../util/Events.js');
+
+module.exports = (client, { d: data }) => {
+  const guild = client.guilds.cache.get(data.guild_id);
+  if (!guild) return;
+
+  const guildScheduledEvent = guild.scheduledEvents._add(data);
+
+  /**
+   * Emitted whenever a guild scheduled event is created.
+   * @event Client#guildScheduledEventCreate
+   * @param {GuildScheduledEvent} guildScheduledEvent The created guild scheduled event
+   */
+  client.emit(Events.GuildScheduledEventCreate, guildScheduledEvent);
 };
