+++ packages/discord.js/src/client/websocket/handlers/THREAD_MEMBER_UPDATE.js
@@ -1,5 +1,25 @@
 'use strict';
 
-module.exports = (client, packet) => {
-  client.actions.ThreadMemberUpdate.handle(packet.d);
+const Events = require('../../../util/Events.js');
+
+module.exports = (client, { d: data }) => {
+  // Discord sends the thread id as id in this object
+  const thread = client.channels.cache.get(data.id);
+  if (!thread) return;
+
+  const member = thread.members.cache.get(data.user_id);
+  if (!member) {
+    thread.members._add(data);
+    return;
+  }
+
+  const old = member._update(data);
+
+  /**
+   * Emitted whenever the client user's thread member is updated.
+   * @event Client#threadMemberUpdate
+   * @param {ThreadMember} oldMember The member before the update
+   * @param {ThreadMember} newMember The member after the update
+   */
+  client.emit(Events.ThreadMemberUpdate, old, member);
 };
